トロフィー機能を実装

+++ db/migrate/20250410153436_drop_skin_care_category.rb
@@ -1,9 +1,11 @@
 class DropSkinCareCategory < ActiveRecord::Migration[7.2]
   def change
-    drop_table :skin_care_categories do |t|
-      t.string :name, null: false
-      t.timestamps
-      t.index :name, unique: true
+    if table_exists?(:skin_care_categories)
+      drop_table :skin_care_categories do |t|
+        t.string :name, null: false
+        t.timestamps
+        t.index :name, unique: true
+      end
     end
   end
 end
